Minimize dependabot requests
This is a static site that is managed on a volunteer basis, as such
I think it is safe to slow dependabot notifications to monthly AND to
group them together into a single check.

If the project gets more complicated we should consider adjusting the
group in particular.

Issue #59 Update dependabot config